Move app URL scheme code into separate file and add /new/note-name/encoded-note-contents route handler.
